refactor: Refactor AspireBudget
Remove module methods out of configuration.rb
