+++ lib/aspire_budget.rb
@@ -8,3 +8,17 @@ require 'worksheets/category_transfers'
 
 require 'models/transaction'
 require 'models/category_transfer'
+
+module AspireBudget
+  def self.configure
+    yield(configuration)
+  end
+
+  def self.configuration
+    @configuration ||= Configuration.new
+  end
+
+  def self.reset!
+    @configuration = Configuration.new
+  end
+end
